# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fernando Portfolio API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/v1/weather/{city}": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Busca clima atual",
                "operationId": "e9ef4278fa7c1e73244e3cd77de70f98",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}